Do not add the resource when morphing a GC_Contact to a normal Contact.
gc_contact.jid and gc_contact.resource are empty in anonymous rooms. In non-anonymous rooms those contain the real jid and resource of the contact.
We must not mix those information with the jid of the contact in the groupchat: We assumed that "test@conference.gajim.org/nick" was the Jid and 'someresource' the resource. So we tried to talk to 'test@conference.gajim.org/nick/someresource'.