Make the port optional in <streamhost/>
XEP-0065 section 9.2 says “The "port" attribute specifies the port to
connect to.  This attribute MAY be present.  The value MUST be a valid
port number in decimal form.  If not specified, the port value is
"1080" (in accordance with RFC 1928).”

This fixes a traceback when a streamhost is advertised with the default
port.