Allow IBB FT to bare JIDs again
Currently works on released gajim, but not on nightly.

This explicit check for resource in the to-jid just returns if there is no
resource. It doesn't even cancel the file transfer or anything, it just
stops... which looks to the user like a file transfer hang.  Removing
the check makes Gajim work as expected.

Closes #8581