Allow preferences window to shrink
Previously, it would always be the size needed to fit the largest
page from the notebook.  This is death on small-screen devices like
the pocketCHIP.

Just adding a ScrolledWindow allows is to get smaller (at least
vertically, which is the main issue).  However, we want to still be a
nice size on large screens.  So, we ask each page in the notebook how
tall it wishes to be, and default the window to the tallest one (plus
the size of the extra controls).  However, if the screen is shorter than
this, we do not allow the default window size to exceed the screen size.

Because of the large amount of re-indentation in the XML file, it is
suggested to view this diff with `-b`