Dont show avatars as Window icon
Avatars are now Surfaces, Gtk.Window has no method yet to add Surfaces
as icons.

Transforming them back to pixbufs is not viable because pixbufs cant hold
the scale information, and we have no way to set it.

This was only used with single chat windows, which is a mode that came
out of style and is probably not used often anymore.

Fixes #8895