MessageWindow: Refactor shortcut handling
AccelGroup/AccelMap is phased out in GTK3.

Use actions instead because they can also have shortcuts. This makes
the code much easier to read because we dont have to deal with keyval and key modifiers.

Handle all actions that are not control specific inside the MessageWindow and
delegate all control specific actions to the controls.