Jingle video: Replace X11 specifics with GTK+
This uses the same mechanism originally added for preferences, so that
video display is handled by a GTK+ widget instead of some X11 dark
magic.

This makes the self video display properly (!) during a Jingle call, but
still with random freezes and no remote video being displayed, those
fixes will be for a latter commit.