Preferences: Add a video preview widget
In order to work on every system Gajim runs on (instead of only X11),
this new widget uses GTK+ Gstreamer elements.

It currently completely breaks the actual Jingle thing, but that was
already pretty broken so will be fixed in a later commit.

It might be nice to think about the way we want to make preferences
interact with a currently-running audio/video session, since it
currently uses a separate pipeline and thus will conflict when e.g. both
will want to access the same camera device, if the OS doesn’t implement
multiplexing (Linux doesn’t).