ChatControlBase: Add focus() method
This method is called by the MessageWindow on the
currently active control, when it gains focus.

This allows the controls to decide by themself what should be focused
inside the control.

For example the GroupchatControl could have the change nickname page
open and would want to focus the nickname entry