Client: Add ClientState.AVAILABLE
When we are connected to the server we need to set some things up
before we send our initial available presence. The new state makes
it possible for other code to difference between when we are connected
and when we are available.

For example other code should not send a presence before we requested
the roster.

Fixes #9983