fix: GroupchatInfo: Don’t set subject when loading from disco info
Disco provides only the subject text not the author and timestamp.
It’s better to set this info separately when we receive the subject
from the MUC. This has the drawback that we can’t show the subject
when we only show the disco (StartChatDialog). But everything else
would bring alot more complexity than it’s worth.

Fixes #11040