imprv: Chatstate: Don’t make too many assumptions
- active does not necessarily mean the user looks at the chat

- gone: In todays chat clients there is no difference between closing a chat
and not looking at it or switching to another chat. In both cases the user gets notified if new
messages arrive, its not clear how this informtation can be important
for our users.

Further for communicating idle states there are more useful mechanisms like
idle presence and presence show states.