imprv: Restore ability to manually disambiguate JID-like addresses
...by storing them in `href` attributes as internal-use
`about:ambiguous-address?a@b.c` URIs.

Storing raw `a@b.c` in `href` wouldn't make a lot of semantic sense, as
that's a relative-path URI with the meaning of "the thing named 'a@b.c'
in the same directory as the current document".  Plus, it would either
entail handling these addresses outside `parse_uri()`, or changing its
purpose of parsing only absolute URIs.