imprv: Notify of new members in private MUCs
Participants joins and leaves are not shown by default,
which is a good thing in a world where a lot of people
are using mobile.

That said, in private groups, knowing if there is a new
member is valuable information, and most chat apps will
inform their users of such events.

This commit implements that feature.