feat: Voice Messages
Single click to open the recording popover. Click + hold to immediately start a recording.

When closing the popover, the user is prompted to either send or to discard the attached recording file.
In case the user decides not to send the file (cancel) the current approach is to start a new recording afterwards. The files remain on disk until Gajim quits.

Whenever a new recording is started, it records into a new file called voice-message-date.m4a.part{i} in a temporary directory, where {i} is an increasing index for the number of resumed recordings. When the recording stops, all voice-message-date.m4a.part{i} files will be merged together into a final file.

The audio preview class has been extended to load audio files on demand after creating an instance.
The audio visualizer has been enriched: Besides drawing a static graph, it offers a treadmill like animation for ongoing records.
To improve visuals in this case, it offers now to not draw gaps between the two displayed channels and to omit normalization to the largest peak displayed as well as raising smaller peaks a bit, which actually sound louder played back than they'd be visually perceived otherwise.