fix: Group chats: Fix resetting settings to default
When using get_contact() on JIDs from the database, we might get a (new)
BareContact back if we haven't seen this JID in our current session.
This can be the case for group chats we've left in previous sessions.
Call get_contact() with groupchat=True to make sure that we always get a
GroupchatContact.

Fixes #11377