fix: VoiceMessageRecorderWidget: Don't fail on switching input device
After switching the input device from autoaudiosrc to something else,
recording would work only after the second attempt.

We want to ensure pipeline is in a state, where we can savely switch.
Furthermore we only ever want to switch, while no recording is in
progress, i.e. we must not set the source element into the playing
state at the end of the switch.